rpy_mimansa_apps_frontend/requirements.txt
asgiref==3.3.4
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==8.1.0
Django==3.2.3
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-api-logger==1.0.6
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1